Fix what a blind review of the LLVM 20 port found
Eight reviewers, one partition each, given the diff, the 20.1.8 headers and
the llvm-project source at both tags, and nothing anyone had said about the
branch. What they found that an audit of suspicions had not was contract
changes: upstream methods the shim still calls by the same name, whose
bodies changed underneath it. Every crash below was reproduced by probe
before it was touched.

Regressions from the port, fixed:

- `FieldDecl::getBitWidthValue` stopped evaluating the width and reads an
  integer cached in a `ConstantExpr` through an unchecked cast. A plain field
  segfaulted and a width installed with `setBitWidth` came back as garbage.
  It and `isZeroLengthBitField` now restate the precondition.
- `NonTypeTemplateParmDecl::getDefaultArgument` returns an empty
  `TemplateArgumentLoc` where it returned a null `Expr*`, and reading the
  expression out of that asserts: a parameter with no default aborted the
  process. The wrapper answers the documented null itself, and the C header
  no longer promises one.
- `Sema::getMoreSpecializedTemplate` gained `RawObj1Ty`/`RawObj2Ty`, which
  the shim filled with null types that clang dereferences; ordering two
  member function templates segfaulted where LLVM 18 worked. The shim and the
  wrapper take upstream's signature now. The object types default to the
  class each template belongs to, as overload resolution passes them, and a
  null one is refused exactly where clang would read it.
- Code generation reads `CompilerInstance::getTimerGroup()` under
  `-ftime-report`, which only `cc1_main` creates; the flag segfaulted
  `create_interpreter` and `create_irgenerator`. `CodeGenOptions` gains
  `getTimePasses`/`setTimePasses`, the drivers create the timer when it is
  set, and `EmitBackendOutput` restates the precondition.
- `Language::CIR` joined the languages `getDefaultLanguageStandard` answers
  with `llvm_unreachable`; two gates still listed only the other two.
- Ten enum mirrors lacked enumerators LLVM 20 appended, so `getCallConv` on
  a `preserve_none` function was `<invalid #22>`. ENUM_SYNC cannot see this
  -- nothing names the missing enumerator -- so each of the ten also gets a
  `switch` over the upstream enum with no default and `-Wswitch` as an error:
  the next appended enumerator stops the build and names itself.
  `CXTemplateName_NameKind` had no ENUM_SYNC table and one enumerator spelled
  without its prefix; both fixed.
- `Decl::getGlobalID` is 64 bits in LLVM 20, the module-file index in the
  upper half, and crossed as `unsigned`. It and the 47 `CreateDeserialized`
  ID parameters are `uint64_t`.
- `getMinGlobalAlign` hard-coded `false` for upstream's new `HasNonWeakDef`,
  the opposite of what clang passes when it has no declaration; it is a
  parameter, defaulting to `true`.

Reachable on main as well, fixed in passing: `CompileDtorCall` on a record
with no definition; a builtin ID past the end of the table, which
`Sema.CreateBuiltin` and fourteen `Builtin::Context` accessors indexed
unchecked -- `getNumBuiltins` reads the two private target tables' sizes so
the gate has an upper bound; `create_interpreter` returning a null
interpreter where construction failed, now an error that names
`create_parser`.

Also: a JLL with no build for the running LLVM is reported as that, rather
than as an `UndefVarError` from inside the JLL. The fabricated
`ModuleConstructorTag` is static_asserted empty; `getMaxBuiltinID` derives
from the sentinel instead of spelling 65533. The generator pins
`LLVM_full_jll` to the LLVM it names its output after: unpinned it installed
23.1.1, and handed a `VersionNumber` Pkg 1.13 fails to set it up, so the pin
is a string. Regenerating against 20.1.8 changed nothing but what the headers
above changed.

Tests. The critics of the test rewrite found what it still could not see:
fourteen more wrapper methods with no caller (a name-level check had missed
per-receiver overloads such as `getSourceRange`); seven `test/types.jl`
assertions turned from `isa Carrier` into `==`, which on carriers is pointer
identity and cannot fail; some thirty-five asm assertions behind guards with
no else; twenty-two `x === nothing || ...`, which pass whichever outcome
clang chose; an MSGuid fixture whose value existed only under LLP64; and
`# shape-only` markers whose reason was false, among them a module's
availability, which its constructor decides. All are pins, relations or both
polarities now, each checked under the x86_64 Linux and Windows triples by
cross-parsing where a target could decide it. The six constant-returning
stubs get a call each again: the constant is this repo's own C, but the call
is what checks the binding still marshals.

CI, packaging and prose:

- deps/build_ci.jl decided whether to rebuild the shim by comparing the time
  of the last ClangExtra commit with the time of the JLL's tag, which says
  nothing about which LLVM that JLL was built for. One with no artifact for
  this platform and LLVM is now never "the JLL's support library", however
  recent its tag.
- The bindings job diffed `lib/` only, while the generator it runs also
  writes fourteen files under `src/`, and `git diff` does not see a new file.
  It fails on anything changed or added under either.
- The bindings and examples jobs get the timeout the test job has.
- deps/ClangExtra/CLAUDE.md describes the exhaustive-switch guard beside
  ENUM_SYNC, and what the bindings job now checks.
- MARSHALLING.md cited `Driver::getLTOMode(true)` and a line of a julia-18
  tag. LLVM 20 spells the offload mode `getOffloadLTOMode()`; the claim that
  only it is uninitialised still holds, and no longer carries a line number.
- examples/06 pointed at "§3" for the act that runs code; it is Act 4.
- The api-coverage skill named the LLVM 18 artifact by hash. It finds the
  LLVM_full artifact for the running Julia's LLVM by version, and no longer
  cites a test file that does not exist.

A hand-built `Module`'s bits. test/CLAUDE.md filed a synthetic module's
availability, its `isPartOfFramework` and the `markUnavailable` transition
under "nothing decides it": reads of memory nobody wrote, because such a module
never had a module map to set those bits. `clang::Module`'s constructor sets
every one of them in its mem-init list. They are equalities now and hold on all
three runners, Windows included, where an earlier disagreement had been taken
as proof of the opposite; `isPartOfFramework` is false for a plain module and
true for one constructed as a framework, the bit being a constructor argument.
The guidance says what the episode taught: "nothing wrote it" is a claim about
a mem-init list and is checked there, not inferred from how the object came to
exist. It also names the second way a value comes back outside its enum, which
this branch met ten times: a mirror that lacks an enumerator upstream appended.
`checkModuleIsAvailable`'s docstring called a hand-built module's availability
host-decided; a module is born available.

+++ deps/ClangExtra/include/clang-ex/AST/CXDeclBase.h
@@ -304,7 +304,7 @@ bool clang_Decl_isFromASTFile(CXDecl D);
 
 // Both IDs live in the words the AST reader writes in front of a deserialized
 // Decl; clang returns 0 for a decl that was parsed instead of deserialized.
-unsigned clang_Decl_getGlobalID(CXDecl D);
+uint64_t clang_Decl_getGlobalID(CXDecl D);
 
 unsigned clang_Decl_getOwningModuleID(CXDecl D);
